Change FinalMark and Grade correct and incorrect to int

+++ src/main/java/com/example/quiz/domain/FinalMark.java
@@ -12,11 +12,11 @@ public class FinalMark {
     this.incorrect = incorrect;
   }
 
-  public long correct() {
+  public int correct() {
     return correct;
   }
 
-  public long incorrect() {
+  public int incorrect() {
     return incorrect;
   }
